requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
astroid==3.3.3
astunparse==1.6.3
attrs==23.2.0
audioread==3.0.1
av==13.0.0
beautifulsoup4==4.12.3
black==24.8.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comtypes==1.4.7
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
decord==0.6.0
dill==0.3.8
dnspython==2.6.1
docutils==0.21.2
facenet-pytorch==2.6.0
Farama-Notifications==0.0.4
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.15.4
flake8==7.1.1
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
fvcore==0.1.5.post20221221
gast==0.6.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
gradio==4.44.0
gradio_client==1.3.0
grpcio==1.66.2
gymnasium==0.29.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
iopath==0.1.10
isort==5.13.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.5.0
Keras-Preprocessing==1.1.2
keyring==25.4.1
kiwisolver==1.4.5
langchain-core==0.3.5
langchain-ollama==0.2.0
langsmith==0.1.128
lazy_loader==0.4
libclang==18.1.1
librosa==0.9.2
lime==0.2.0.1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
mtcnn==0.1.1
mypy==1.11.2
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.3
nh3==0.2.18
numba==0.60.0
numpy==1.26.4
ollama==0.3.3
omegaconf==2.2.2
onnx==1.17.0
onnx2pytorch==0.5.0
opencv-python==4.5.5.64
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.12.1
orjson==3.10.7
outcome==1.3.0.post0
packaging==24.1
pandas==2.1.1
parameterized==0.9.0
pathspec==0.12.1
pillow==10.2.0
pkginfo==1.10.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
portalocker==2.10.1
proglog==0.1.10
protobuf==4.25.5
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
PyAudio==0.2.14
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
pyflakes==3.2.0
pygame==2.6.0
Pygments==2.18.0
pylint==3.3.0
pymongo==4.8.0
Pympler==1.1
pyparsing==3.1.2
pypiwin32==223
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorchvideo==0.1.5
pyttsx3==2.91
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.1
readme_renderer==44.0
referencing==0.35.1
regex==2024.9.11
reportlab==4.2.2
requests==2.32.3
requests-toolbelt==1.0.0
resampy==0.4.3
rfc3986==2.0.0
rich==13.9.1
rpds-py==0.19.1
ruff==0.6.5
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.0
seaborn==0.13.2
selenium==4.23.1
semantic-version==2.10.0
semver==3.0.2
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
SpeechRecognition==3.10.4
stable_baselines3==2.3.2
starlette==0.38.5
streamlit==1.10.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.9.20
tokenizers==0.20.0
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.1
trio==0.26.2
trio-websocket==0.11.1
twine==5.1.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ultralytics-thop==2.0.0
urllib3==2.2.3
uvicorn==0.30.6
validators==0.34.0
watchdog==4.0.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
yacs==0.1.8
zipp==3.20.2
